add cost as a property new-rental.js's 'save' action
Note: Make sure that the rental cost is being saved as an Integer, and
not a string in order to sort correctly. If numerical values are saved
as Strings JavaScript, regardless of framework, will think "76" is
larger than "1000" because the first character is larger. You can
ensure all new rental costs will be saved as Integers by simply parsing
the value when you gather the params: